# Conflicts:
#	indra/newview/llinventorybridge.cpp
#	indra/newview/llinventoryfilter.cpp
#	indra/newview/llpanelmaininventory.cpp
#	indra/newview/llpanelobjectinventory.cpp
#	indra/newview/llviewerregion.cpp
#	indra/newview/llworldmapview.cpp
#	indra/newview/skins/default/xui/en/floater_tools.xml
#	indra/newview/skins/default/xui/en/menu_viewer.xml

+++ indra/newview/llviewertexteditor.cpp
@@ -586,7 +586,7 @@ S32 LLEmbeddedItems::getIndexFromEmbeddedChar(llwchar wch)
     }
     else
     {
-        LL_WARNS() << "Embedded char " << wch << " not found, using 0" << LL_ENDL;
+        LL_WARNS() << "Embedded char " << (int)wch << " not found, using 0" << LL_ENDL;
         return 0;
     }
 }
